+++ playground/App.vue
@@ -29,6 +29,7 @@
     <TestWasm />
   </Suspense>
   <TestScriptSetupStyleVars msg="Test message" />
+  <TestSyntax/>
 </template>
 
 <script>
@@ -58,8 +59,9 @@ import TestDynamicImport from './dynamic-import/TestDynamicImport.vue'
 import TestWebWorker from './worker/TestWorker.vue'
 import TestWasm from './wasm/TestWasm.vue'
 import TestScriptSetupStyleVars from './script-setup/TestScriptSetupStyleVars.vue'
+import TestSyntax from './TestSyntax.vue'
 
-export default {
+const App = {
   components: {
     TestEnv,
     TestModuleResolve,
@@ -86,7 +88,9 @@ export default {
     TestDynamicImport,
     TestWebWorker,
     TestWasm,
-    TestScriptSetupStyleVars
+    TestScriptSetupStyleVars,
+    TestSyntax
   }
 }
+export { App as default}
 </script>